CONTRIBUTING: Don't tell people to make clear pull requests
I think this sort of generic hygiene advice lowers the signal/noise
for this file, since I'm more interested in what this project wants
that other projects may not want, and everybody wants well-described
PRs.  And I suspect that folks who are inclined to post
poorly-described PRs are less likely to be reading CONTRIBUTING in
detail anyway ;).

Signed-off-by: W. Trevor King <wking@tremily.us>