MAINTAINERS: Explain the format and reference other projects
Currently the OCI seems to have one repository per Project.  A few
repositories are not official projects; ocitools is just becoming
official [1] and artwork [2] may never be a Project.  But I expect
that going forward, any opencontainers/ repos that have a MAINTAINERS
file will be official Projects, so the union rule here should be
both correct and easy to maintain.

The file is not Markdown, so I've gone with my usual reference-style
link instead of a Markdown reference-style link.

[1]: https://github.com/opencontainers/tob/pull/18
     Subject: Create runtime-tools and image-tools projects
[2]: https://github.com/opencontainers/artwork

Signed-off-by: W. Trevor King <wking@tremily.us>