CONTRIBUTING: Code of conduct, meetings, mailing list, and IRC
So far dev@ volume has been low enough that we haven't needed
per-project lists [1], so we can add it in this central location.  In
the event that we do split lists later, the reference-style link makes
it easy for projects to update the target without touching the line
with the link text.  The meetings and IRC channel are in the same boat
(not enought traffic to be worth splitting).  The content I'm adding
here is adapted from [2,3].  As long as the meeting is shared, linking
the runtime-spec maintained meeting.ics seemed easier than keeping a
parallel copy here.

The CoC titles itself "OpenContainers Code of Conduct", but [4] has:

  Connecting with the Open Container Initiative community

And I feel like I've been seeing the singular, spaced form more frequently.

While we're adding reference-style links, convert the DCO link as well
to make the text easier to read.

[1]: https://groups.google.com/a/opencontainers.org/d/msg/dev/Ctw-fcO1IuA/7vrr4YiyDgAJ
     Subject: Re: splitting this list into image-spec-dev, runtime-spec-dev, etc
     Date: Wed, 20 Apr 2016 14:45:55 -0400
     Message-ID: <CAEXjzRug6uL_1eLHyxUzN+3AchXq6KVtV6KmgxRj0Gu+krfAcg@mail.gmail.com>
[2]: https://github.com/opencontainers/runtime-spec/blob/7a36e7ed86ee3b4c6dbcdbd02052ec1ef6787c3c/README.md#contributing
[3]: https://github.com/opencontainers/runtime-spec/blob/1077f05ae7b05ce79159e66c25f8ff80f58ffedb/README.md#meetings
[4]: https://www.opencontainers.org/community

Signed-off-by: W. Trevor King <wking@tremily.us>