CONTRIBUTING: Remove gofmt advice
Not all OCI Projects are primarily Go, and filling in off-topic
formatting advice doesn't seem useful.  One approach to this would be
to require projects to support 'make fmt' to handle any
auto-formatting.  But an easier approach is to keep the document
generic in this repository [1], and allow downstream projects to add
their own project-specific content as they see fit.  I don't expect a
lot of churn in this document, so there shouldn't be many conflicts
between those downstream changes and further project-template
development.

[1]: https://github.com/opencontainers/project-template/issues/4#issuecomment-221081999

Signed-off-by: W. Trevor King <wking@tremily.us>