MAINTAINERS_GUIDE: Replace Chief Maintainer with GOVERNANCE
Hooray institutions ;).  Because GOVERNANCE is already making
decisions with a 2/3 vote, there's no reason to appeal to the TOB (the
old 2/3 vote for appeal is now sufficient for making the decision
outright).

All current OCI Projects have adopted the GOVERNANCE docs [1]
(although runC has yet to actually merge them into its repository) so
I think this approach is portable while the Chief Maintainer approach
was not [2].  Taking the runC maintainer subset of that vote (just to
be sure the doc applies to runC):

+7: Aleksa Sarai, Alexander Morozov, Daniel Dao, Mrunal Patel, Qiang
    Huang, Rohit Jnagal, Victor Marmol
-0
#2: Andrey Vagin, Michael Crosby

and 7/9 > 2/3.

This also avoids the strange behavior where a 2/3 vote of maintainers
could approve a new maintainer, the Chief Maintainer could veto, and
the same 2/3 vote could appeal that veto to the TOB.  And it's nice to
have a single set of rules for project-management issues, and not a
five "business days" window for new maintainers one-week window for
other management issues.  The ten-day window for maintainer removal is
now a shorter seven, but with the call for earlier private discussion
I don't think it's worth special-casing just to get an extra three
days.

Also:

* Remove "across the maintainers of the project".  "respect across"
  seemed awkward ("respect between" is closer but still not quite
  right).  In any case, the next sentence makes it clear with "trust
  one another", so I think the bit I removed was superfluous.
* Replace "depend on and trust" with "depend on", because building
  trust was already mentioned in that sentence, and I don't see any
  semantic distiction between "depend on and trust" and "depend on".
* Replace "make decisions" with "act".  Same meaning, fewer letters ;).
* Adjust the paragraphs I touched to the README's recommended one line
  per sentence.
* Fixed "point of views" -> "points of view".

[1]: https://groups.google.com/a/opencontainers.org/d/msg/dev/x-Oh3PDz1Y8/q7t2IseVAwAJ
     Subject: [project-template adopted]: Merge 56abe122 (+13 -0 #5)
     Date: Wed, 20 Jul 2016 16:51:58 +0000
     Message-ID: <CAD2oYtPwMcF__WD32cV6dHgHt8=F6qFw+XFGw4iQK9LGi_QWsg@mail.gmail.com>
[2]: https://github.com/opencontainers/runtime-spec/pull/420#discussion_r61972504
     Subject: Update maintainers and contributors guides

Signed-off-by: W. Trevor King <wking@tremily.us>