CONTRIBUTING: Remove docs/README.md reference
Current ocitools docs for the man pages are under man/. No docs for
building them yet beyond the Makefile rule, and that may be all we
need ;).  Because there is no downstream consensus on doc locations,
remove the reference.  Downstream repositories can add it (or similar
content) back in as they see fit [1].  I don't expect a lot of churn
in this document, so there shouldn't be many conflicts between those
downstream changes and further project-template development.

[1]: https://github.com/opencontainers/project-template/issues/4#issuecomment-221081999

Signed-off-by: W. Trevor King <wking@tremily.us>