GOVERNANCE: Update project list to include *-tools
Catching up with opencontainers/tob@729e1ecf (Merge pull request #18
from RobDolinMS/master, 2016-09-18).  I'd rather drop the
parenthetical entirely and link to a place that listed OCI Projects,
but we don't have a canonical target for that yet
(opencontainers/tob#2) and the current closest instance seems to be
the GitHub section in [1] (which doesn't have the "OCI Project"
words).

[1]: https://www.opencontainers.org/community

Signed-off-by: W. Trevor King <wking@tremily.us>