GOVERNANCE: Update project list to include go-digest
Catching up with opencontainers/tob@ce087c84 9Merge pull request #22
from opencontainers/digest-proposal, 2017-01-05).  I'd still rather
drop the parenthetical entirely and link to a place that listed OCI
Projects, but we don't have a canonical target for that yet
(opencontainers/tob#2) and the current closest instance seems to be
the GitHub section in [1] (which doesn't have the "OCI Project"
words).

[1]: https://www.opencontainers.org/community

Signed-off-by: W. Trevor King <wking@tremily.us>