.pullapprove.yml: Add distribution-spec-maintainers
Catching up with [1].

Also mention the new project in GOVERNANCE.md.  I'd still rather drop
the parenthetical entirely and link to a place that listed OCI
Projects, but we don't have a canonical target for that yet
(opencontainers/tob#2) and the current closest instance seems to be
the GitHub section in [2] (which doesn't have the "OCI Project"
words).

[1]: https://github.com/opencontainers/tob/pull/35
[2]: https://www.opencontainers.org/community

Signed-off-by: W. Trevor King <wking@tremily.us>