CONTRIBUTING: Lead off with the security section
This reduces the chances that a skimmer reads "Issues are used for
bugs..." and files an issue without scrolling down to see the security
section.

I've also added links back to the security section from the all other
paragraphs that suggest using other channels (issues, pull requests,
IRC, ...), so that no matter which anchor a skimmer starts at, they
cannot miss the fact that we have a special disclosure procedure for
security issues.  If we are confident that readers interested in
submitting a security issue will not start at a mid-document anchor
without reading the beginning of this file, we could drop some or all
of these back-refs.

Also adjust the lines I touch to use one line per sentence, since
that's the standard in other OCI Projects (like the runtime spec [1]).

[1]: https://github.com/opencontainers/runtime-spec/blob/v1.0.1/style.md#one-sentence-per-line

Signed-off-by: W. Trevor King <wking@tremily.us>